+++ lib/karafka-core.rb
@@ -23,7 +23,6 @@
   karafka/core/contractable/rule
 
   karafka/core/patches/rdkafka/bindings
-  karafka/core/patches/rdkafka/producer
 ].each { |dependency| require dependency }
 
 # Karafka framework main namespace
@@ -31,5 +30,4 @@ module Karafka
 end
 
 # Patch rdkafka
-::Rdkafka::Producer.include(::Karafka::Core::Patches::Rdkafka::Producer)
 ::Rdkafka::Bindings.include(::Karafka::Core::Patches::Rdkafka::Bindings)